requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.2.0
python-crontab==2.6.0
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
typing-extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.1
zipp==3.8.1
